backend/src/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
# packages = [{include = "ptm_core"}]

[tool.poetry.scripts]
start = "app.api.main:start"
generate_predefined_data = "app.db.data.generators.generate_predefined_data:generate"
insert_predefined_data = "app.db.database:insert_predefined_data"

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.27.0"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.5"
autopep8 = "^2.0.1"
isort = "^5.12.0"
alembic = "^1.9.2"
faker = "^17.0.0"


[tool.poetry.group.dev.dependencies]
alembic = "^1.9.1"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
